+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+next version
+------------
+
+* Rename zero_order_absorption to set_zero_order_absorption
+
 0.28.0 (2021-08-24)
 -------------------